+++ packages/backend/src/App.ts
@@ -2,16 +2,9 @@ import { LightdashMode, SessionUser } from '@lightdash/common';
 import { NodeSDK } from '@opentelemetry/sdk-node';
 import * as Sentry from '@sentry/node';
 import { nodeProfilingIntegration } from '@sentry/profiling-node';
-import { SamplingContext } from '@sentry/types';
 import flash from 'connect-flash';
 import connectSessionKnex from 'connect-session-knex';
-import express, {
-    Express,
-    NextFunction,
-    Request,
-    RequestHandler,
-    Response,
-} from 'express';
+import express, { Express, NextFunction, Request, Response } from 'express';
 import expressSession from 'express-session';
 import expressStaticGzip from 'express-static-gzip';
 import helmet from 'helmet';
@@ -217,6 +210,9 @@ export default class App {
     }
 
     async start() {
+        // NOTE: Sentry must be initialized as soon as possible before any relevant initialization - express, knex, etc.
+        this.initSentry();
+
         // @ts-ignore
         // eslint-disable-next-line no-extend-native, func-names
         BigInt.prototype.toJSON = function () {
@@ -233,9 +229,6 @@ export default class App {
         this.initSlack(expressApp).catch((e) => {
             Logger.error('Error starting slack bot', e);
         });
-
-        // NOTE: Sentry must be initialized before any handlers or middleware that should be traced
-        this.initSentry(expressApp);
 
         // Load Lightdash middleware/routes last
         await this.initExpress(expressApp);
@@ -275,11 +268,50 @@ export default class App {
             Logger.warn('Invalid security report URI', e);
         }
 
+        const contentSecurityPolicyAllowedDomains: string[] = [
+            'https://*.sentry.io',
+            'https://analytics.lightdash.com',
+            'https://*.usepylon.com',
+            'https://*.headwayapp.co',
+            'https://headway-widget.net',
+            'https://*.posthog.com',
+            'https://*.intercom.com',
+            'https://*.intercom.io',
+            'https://*.intercomcdn.com',
+            'https://*.rudderlabs.com',
+            'https://www.googleapis.com',
+            'https://apis.google.com',
+            'https://accounts.google.com',
+            ...this.lightdashConfig.security.contentSecurityPolicy
+                .allowedDomains,
+        ];
+
         expressApp.use(
             helmet({
                 contentSecurityPolicy: {
                     directives: {
-                        'default-src': ["'self'", 'sentry.io'],
+                        'default-src': [
+                            "'self'",
+                            ...contentSecurityPolicyAllowedDomains,
+                        ],
+                        'img-src': ["'self'", 'data:', 'https://*'],
+                        'frame-src': ["'self'", 'https://*'],
+                        'frame-ancestors': ["'self'", 'https://*'],
+                        'worker-src': [
+                            "'self'",
+                            'blob:',
+                            ...contentSecurityPolicyAllowedDomains,
+                        ],
+                        'script-src': [
+                            "'self'",
+                            "'unsafe-eval'",
+                            ...contentSecurityPolicyAllowedDomains,
+                        ],
+                        'script-src-elem': [
+                            "'self'",
+                            "'unsafe-inline'",
+                            ...contentSecurityPolicyAllowedDomains,
+                        ],
                         'report-uri': reportUri ? [reportUri.href] : [],
                     },
                     reportOnly: true,
@@ -293,6 +325,7 @@ export default class App {
                     policy: 'strict-origin-when-cross-origin',
                 },
                 noSniff: true,
+                xFrameOptions: false,
             }),
         );
 
@@ -418,7 +451,7 @@ export default class App {
         });
 
         // Errors
-        expressApp.use(Sentry.Handlers.errorHandler()); // The Sentry error handler must be before any other error middleware and after all controllers
+        Sentry.setupExpressErrorHandler(expressApp);
         expressApp.use(
             (error: Error, req: Request, res: Response, _: NextFunction) => {
                 const errorResponse = errorHandler(error);
@@ -519,7 +552,7 @@ export default class App {
         await slackBot.start(expressApp);
     }
 
-    private initSentry(expressApp: Express) {
+    private initSentry() {
         Sentry.init({
             release: VERSION,
             dsn: this.lightdashConfig.sentry.backend.dsn,
@@ -528,13 +561,10 @@ export default class App {
                     ? 'development'
                     : this.lightdashConfig.mode,
             integrations: [
-                new Sentry.Integrations.Http({ tracing: true }),
-                new Sentry.Integrations.Express({
-                    app: expressApp,
-                }),
-                new Sentry.Integrations.Postgres({ usePgNative: true }),
+                Sentry.httpIntegration({ breadcrumbs: true }),
+                Sentry.expressIntegration(),
+                Sentry.postgresIntegration(),
                 nodeProfilingIntegration(),
-                ...Sentry.autoDiscoverNodePerformanceMonitoringIntegrations(),
                 ...(this.lightdashConfig.sentry.anr.enabled
                     ? [
                           Sentry.anrIntegration({
@@ -550,7 +580,7 @@ export default class App {
                     : []),
             ],
             ignoreErrors: ['WarehouseQueryError', 'FieldReferenceError'],
-            tracesSampler: (context: SamplingContext): boolean | number => {
+            tracesSampler: (context) => {
                 if (
                     context.request?.url?.endsWith('/status') ||
                     context.request?.url?.endsWith('/health') ||
@@ -581,17 +611,6 @@ export default class App {
                 return breadcrumb;
             },
         });
-        expressApp.use(
-            Sentry.Handlers.requestHandler({
-                user: [
-                    'userUuid',
-                    'organizationUuid',
-                    'organizationName',
-                    'email',
-                ],
-            }) as RequestHandler,
-        );
-        expressApp.use(Sentry.Handlers.tracingHandler());
 
         // Set k8s tags for Sentry
         Sentry.setTags({