+++ packages/common/src/index.ts
@@ -78,7 +78,7 @@ import {
     type OrganizationProject,
     type UpdateAllowedEmailDomains,
 } from './types/organization';
-import { type PinnedItems } from './types/pinning';
+import { type ApiTogglePinnedItem, type PinnedItems } from './types/pinning';
 import { type ProjectGroupAccess } from './types/projectGroupAccess';
 import { type ProjectMemberRole } from './types/projectMemberRole';
 import {
@@ -641,7 +641,8 @@ type ApiResults =
     | ApiAiGetDashboardSummaryResponse['results']
     | ApiCatalogMetadataResults
     | ApiCatalogAnalyticsResults
-    | ApiPromotionChangesResponse['results'];
+    | ApiPromotionChangesResponse['results']
+    | ApiTogglePinnedItem['results'];
 
 export type ApiResponse<T extends ApiResults = ApiResults> = {
     status: 'ok';